Remove obsolete color-theme
Since Emacs 24.1, released in 2012, Emacs comes with support
for themes, making the older `color-theme' package obsolete
and incompatible.

`color-theme' is one of the package that Melpa gets from the
Emacsorphanage, which I update semi-automatically and not very
often from the upstream non-git/hg repository.

This particular repository was not updated in ten years, which
isn't surprising since it has been obsolete for six of those.

I have stopped updating this mirror now and its recipe could be
updated to fetch "updates" from the Emacsattic instead.  But we
might as well take this opportunity and remove `color-theme'
and all old-school themes.

There are six theme packages whose repositories contain old-school
themes beside modern themes.  There are no old-school-only theme
packages.

The recipes of three of those already limit the Melpa package to
the modern theme.  The other tree, two of which by Steve, didn't
do so until this commit.

I think these packages should also be renamed from `color-theme-*'
to `*-theme(s)'.  I would also recommend that the `color-theme-*'
files be removed from all repositories (or moved to a branch named
"obsolete" or something like that).