[WIP] Rename theme packages
!! Due to issues mentioned below, this cannot actually be merged yet.

These packages should be named after the new-style themes they provide
instead of after the old-style themes their repositories still contain.

I would encourage you to also remove the obsolete themes from the
repositories, but that can happen in a second step.

The `solarized-theme.el' library of package `sellout-solarized-theme'
conflicts with the library by the same name of the `solarized-theme'
package.  I think it should be renamed to `sellout-solarized-theme.el'
as is done for `sanityinc-solarized-theme'.